+++ internal/modules/community/infrastructure/consumer/ConsumerGroupMember.go
@@ -210,16 +210,6 @@ func (c *ConsumerGroupMember) handleDeletedGroupMember(ctx context.Context, even
 	if data == nil {
 		return kafka.NewNonRetryableError(fmt.Errorf("payload is nil for event %s", event.ID))
 	}
-	dataaction, err := c.groupMemberRepo.GetGroupMemberByUserIDAndGroupID(ctx, data.UserActionID, data.GroupID)
-	if err != nil {
-		return fmt.Errorf("failed to fetch action user group member from repository for event %s: %w", event.ID, err)
-	}
-	if dataaction == nil {
-		return kafka.NewNonRetryableError(fmt.Errorf("action user group member with user_id %s and group_id %s not found for event %s", data.UserActionID, data.GroupID, event.ID))
-	}
-	if dataaction.Role != sharedEnums.RoleTypeAdmin {
-		return kafka.NewNonRetryableError(fmt.Errorf("user with ID %s does not have permission to update group member in group %s for event %s", data.UserActionID, data.GroupID, event.ID))
-	}
 	if data.DeleteALL == true {
 
 		err = c.groupMemberRepo.DeleteGroupMember(ctx, data.GroupID)
@@ -227,6 +217,16 @@ func (c *ConsumerGroupMember) handleDeletedGroupMember(ctx context.Context, even
 			return fmt.Errorf("failed to delete all group members by group_id in repository for event %s: %w", event.ID, err)
 		}
 	} else {
+		dataaction, err := c.groupMemberRepo.GetGroupMemberByUserIDAndGroupID(ctx, data.UserActionID, data.GroupID)
+		if err != nil {
+			return fmt.Errorf("failed to fetch action user group member from repository for event %s: %w", event.ID, err)
+		}
+		if dataaction == nil {
+			return kafka.NewNonRetryableError(fmt.Errorf("action user group member with user_id %s and group_id %s not found for event %s", data.UserActionID, data.GroupID, event.ID))
+		}
+		if dataaction.Role != sharedEnums.RoleTypeAdmin {
+			return kafka.NewNonRetryableError(fmt.Errorf("user with ID %s does not have permission to update group member in group %s for event %s", data.UserActionID, data.GroupID, event.ID))
+		}
 		err = c.groupMemberRepo.DeleteGroupMemberByUserIDAndGroupID(ctx, data.UserID, data.GroupID)
 		if err != nil {
 			return fmt.Errorf("failed to delete group member in repository for event %s: %w", event.ID, err)
